Edu-CirlceUp/requirements.txt
# EduCircle AI Service — requirements.txt
# Using keras instead of torch

# ── Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
httpx==0.27.0

# ── Database
sqlalchemy==2.0.30
psycopg[binary]==3.1.19
redis==5.0.4
alembic==1.13.1

# ── LLM / RAG
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
faiss-cpu==1.8.0
sentence-transformers==3.0.1
ollama==0.2.1

# ── ML — Keras (NOT torch)
tensorflow==2.16.1     # Keras 3 backend
keras==3.3.3
numpy==1.26.4
scikit-learn==1.5.0

# ── NLP
transformers==4.41.2   # HuggingFace (translation, sentiment)
spacy==3.7.4           # entity extraction for mind maps
textstat==0.7.3        # readability scoring
yake==0.4.8            # keyword extraction

# ── Speech
openai-whisper==20231117   # STT
gTTS==2.5.1                # TTS

# ── File Parsing
pypdf==4.2.0
python-docx==1.1.2
python-pptx==0.6.23
pillow==10.3.0
pytesseract==0.3.10    # OCR for handwritten notes

# ── MinIO
minio==7.2.7

# ── Utilities
pydantic==2.7.3
pydantic-settings==2.3.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
requests==2.32.3
aiofiles==23.2.1